Renamed source folder to correct platform
