+++ src/Scenes_FileDrops.ts
@@ -11,8 +11,8 @@ import * as Grezzo3DS from './oot3d/scenes';
 import * as NNS_G3D from './nns_g3d/scenes';
 import * as J3D from './j3d/scenes';
 import * as CTR_H3D from './Common/CTR_H3D/H3D';
-import * as GC_PVRT from './Common/GC/PVRT';
-import * as JSR from './Common/GC/jsr';
+import * as DC_PVRT from './Common/DC/PVRT';
+import * as JSR from './Common/DC/jsr';
 import * as RRES from './rres/scenes';
 import * as PaperMarioTTYD from './PaperMarioTTYD/Scenes_PaperMarioTTYD';
 import * as JPAExplorer from './interactive_examples/JPAExplorer';
@@ -107,7 +107,7 @@ export async function createSceneFromFiles(context: SceneContext, buffers: Named
 
     if (buffer.name.toLowerCase().endsWith('.pvr')) {
         // Load texture data
-        let image = GC_PVRT.parse(buffer, buffer.name);
+        let image = DC_PVRT.parse(buffer, buffer.name);
 
         // Create faux scene
         const jsrScene = new JetSetRadioScene();
